Adding HomePage and moving webgazer to a separate page.
